bitvec: Optimize performance
This patch changes the way that we use BitVec to optimize the packet
forwarding performance.  Instead of converting u64 slice to BitVec
whenever we need to access the flow map of a packet, this patch only do
the coversion once per packet.  With this patch the packet forwarding
rate increase from ~75 KPPS to 120 KPPS.